WIP remove external auth view
