+++ crates/api_crud/src/site/read.rs
@@ -5,10 +5,11 @@ use lemmy_api_common::{
 };
 use lemmy_db_schema::source::{
   actor_language::{LocalUserLanguage, SiteLanguage},
+  external_auth::ExternalAuth,
   language::Language,
   tagline::Tagline,
 };
-use lemmy_db_views::structs::{CustomEmojiView, ExternalAuthView, LocalUserView, SiteView};
+use lemmy_db_views::structs::{CustomEmojiView, LocalUserView, SiteView};
 use lemmy_db_views_actor::structs::{
   CommunityBlockView,
   CommunityFollowerView,
@@ -80,7 +81,7 @@ pub async fn get_site(
   let taglines = Tagline::get_all(&mut context.pool(), site_view.local_site.id).await?;
   let custom_emojis =
     CustomEmojiView::get_all(&mut context.pool(), site_view.local_site.id).await?;
-  let external_auths = ExternalAuthView::get_all(&mut context.pool()).await?;
+  let external_auths = ExternalAuth::get_all(&mut context.pool()).await?;
 
   Ok(Json(GetSiteResponse {
     site_view,
