+++ crates/db_views/src/structs.rs
@@ -185,11 +185,3 @@ pub struct CustomEmojiView {
   pub custom_emoji: CustomEmoji,
   pub keywords: Vec<CustomEmojiKeyword>,
 }
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[cfg_attr(feature = "full", derive(TS))]
-#[cfg_attr(feature = "full", ts(export))]
-/// An external auth view.
-pub struct ExternalAuthView {
-  pub external_auth: ExternalAuth,
-}
